requirements.txt
pyqt5
pyscreenshot
pillow